moved types to its own file
