+++ utils/parse.ts
@@ -6,17 +6,8 @@ import {
   ICalRepeatingOptions,
   ICalWeekday,
 } from "ical-generator";
+import type { IntermediateEventData } from "./types";
 import { endDate, nowSem } from "./constants";
-
-export interface IntermediateEventData {
-  prof?: string;
-  section?: string;
-  subject?: string;
-  location: string;
-  weekdays: ICalWeekday[];
-  start: Dayjs;
-  end: Dayjs;
-}
 
 function weekdayStartFromPrefix(prefix: string): {
   weekdays: ICalWeekday[];
